ref : 정기 제안 컨트롤러 통합
